Fix compiler, cppcheck and clang-tidy findings; clean up includes
Lint profile: GCC -Wall -Wextra -Wshadow on all three boards,
cppcheck via pio check, clang-tidy with the new .clang-tidy (run on
the host with the native fakes). All three now report zero findings
in include/ and src/.

Bugs found along the way:
- WebDebugServer::extractJsonNumber only accepted "key":value with no
  spaces, so the documented {"mode": 1} was rejected and setpid with
  spaces silently kept the old gains; numbers in exponent form
  (JSON.stringify(1e-7)) were read as their mantissa. Whitespace and
  full JSON number syntax are now accepted.
- MagnetometerBase::calibrate() with no successful read stored
  INT16_MAX/INT16_MIN-derived offsets in NVS as a valid calibration;
  it now keeps the previous calibration.
- UbloxM10_Gps::printStatus() reported available=YES after the GPS
  timed out; it now uses isAvailable().
- IBusReceiver dropped a real frame when a stray 0x20 was followed by
  the frame's own 0x20 0x40 header.
- OledDisplay loop-stats line could be truncated (32-byte buffer for
  up to 34 bytes).
- DebugLogger kept GPS coordinates in float (~0.4 m steps); the
  change filter now stores double.

Warnings fixed:
- constructor parameters and locals shadowing members, globals and
  Arduino's byte type (renamed parameters; main.cpp's gps pointer is
  now gpsReceiver);
- memset on structs with floats and uninitialised members replaced by
  value-initialisation;
- implicit double arithmetic from Arduino's DEG_TO_RAD/RAD_TO_DEG in
  float code (software double on ESP32; the attitude filter runs it at
  500 Hz);
- missing explicit/override, AutopilotMode as uint8_t enum, const
  pointers, duplicated branch in TakeoffSequencer, always-true
  condition in the log menu (key mapping now derived from channelKey);
- Esp32Board::hasDisplayBus() exists only where the second I2C bus
  does.

Includes: every header includes what it uses directly
(misc-include-cleaner is clean and each header compiles on its own);
umbrella and API headers carry IWYU export pragmas; include blocks are
sorted and grouped consistently.

Reference docs updated for the changed signatures and behaviour.

+++ include/autopilot/feedback/FeedbackMath.h
@@ -20,7 +20,7 @@ namespace FeedbackMath
 
     inline int8_t signOf(float x)
     {
-        return x > 0.0f ? 1 : (x < 0.0f ? -1 : 0);
+        return static_cast<int8_t>(x > 0.0f ? 1 : (x < 0.0f ? -1 : 0));
     }
 
     inline float clampAbs(float x, float limit)
